requirements.txt
sentencepiece==0.1.97
pytorch-lightning==2.1.2
transformers==4.36.2
datasets==2.16.1
pandas==2.1.4
protobuf==4.25.1